# internal/features/tasks/transport/http/delete_task.go
package tasks_transport_http

import (
	"net/http"

	core_logger "github.com/ReGeC/golang-todoapp/internal/core/logger"
	core_http_request "github.com/ReGeC/golang-todoapp/internal/core/transport/http/request"
	core_http_response "github.com/ReGeC/golang-todoapp/internal/core/transport/http/response"
)

// DeleteTask 	godoc
// @Summary 	Удаление задачи
// @Description Удаление конкретной задачи по её id в системе
// @Tags 		tasks
// @Param 		id 		path int true "id удаляемой задачи"
// @Success 	204 	"Успешное удаление задачи"
// @Failure 	400 	{object} core_http_response.ErrorResponse "Bad request"
// @Failure 	404 	{object} core_http_response.ErrorResponse "Task not found"
// @Failure 	500 	{object} core_http_response.ErrorResponse "Internal Server Error"
// @Router 		/tasks/{id} [delete]
func (h *TasksHTTPHandler) DeleteTask(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	taskID, err := core_http_request.GetIntPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to get 'TaskID path value",
		)

		return
	}

	err = h.tasksService.DeleteTask(ctx, taskID)
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to delete task",
		)

		return
	}

	responseHandler.NoContentResponse()
}