# internal/features/users/transport/http/delete_user.go
package users_transport_http

import (
	"net/http"

	core_logger "github.com/ReGeC/golang-todoapp/internal/core/logger"
	core_http_request "github.com/ReGeC/golang-todoapp/internal/core/transport/http/request"
	core_http_response "github.com/ReGeC/golang-todoapp/internal/core/transport/http/response"
)

// DeleteUser 	godoc
// @Summary 	Удаление пользователя
// @Description Удаление конкретного пользователя по его id в системе
// @Tags 		users
// @Param 		id 		path int true "id удаляемого пользователя"
// @Success 	204 	"Успешное удаление пользователя"
// @Failure 	400 	{object} core_http_response.ErrorResponse "Bad request"
// @Failure 	404 	{object} core_http_response.ErrorResponse "User not found"
// @Failure 	500 	{object} core_http_response.ErrorResponse "Internal Server Error"
// @Router 		/users/{id} [delete]
func (h *UsersHTTPHandler) DeleteUser(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	userID, err := core_http_request.GetIntPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to get userID path value",
		)
	}

	if err := h.usersService.DeleteUser(ctx, userID); err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to delete user",
		)

		return
	}

	responseHandler.NoContentResponse()
}